Move Builder to a standalone component and update lead magnet function
Extracted the Builder functionality from `page.tsx` to a new file `Builder.tsx` for better modularization. Additionally, modified `updateLeadMagnet` to include user ID for more precise data handling. This restructuring helps improve the maintainability and readability of the codebase.
