+++ apps/web/package.json
@@ -13,14 +13,15 @@
   },
   "dependencies": {
     "@auth/prisma-adapter": "^2.5.3",
+    "@smartleadmagnet/services": "*",
     "@smartleadmagnet/ui": "*",
+    "axios": "^1.7.7",
     "next": "^14.2.14",
     "next-auth": "^5.0.0-beta.22",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "tailwind-merge": "^2.5.2",
-    "unique-names-generator": "^4.7.1",
-    "@smartleadmagnet/services": "*"
+    "unique-names-generator": "^4.7.1"
   },
   "devDependencies": {
     "@cloudflare/next-on-pages": "^1.13.5",
