+++ apps/web/src/components/User.tsx
@@ -1,0 +1,57 @@
+import { Button } from '@smartleadmagnet/ui/components/ui/button';
+import Image from 'next/image';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger
+} from '@smartleadmagnet/ui/components/ui/dropdown-menu';
+import { getSessionUser } from '@/services/user';
+import React from 'react';
+import { signOut } from '@/lib/auth';
+
+export async function User() {
+  let user = await getSessionUser();
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button
+          variant="outline"
+          size="icon"
+          className="overflow-hidden rounded-full"
+        >
+          <Image
+            src={user?.image ?? '/placeholder-user.jpg'}
+            width={36}
+            height={36}
+            alt="Avatar"
+            className="overflow-hidden rounded-full"
+          />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="end">
+        <DropdownMenuLabel>My Account</DropdownMenuLabel>
+        <DropdownMenuSeparator />
+        <DropdownMenuItem>Settings</DropdownMenuItem>
+        <DropdownMenuItem>Support</DropdownMenuItem>
+        <DropdownMenuSeparator />
+        <DropdownMenuItem>
+          <form
+            action={async () => {
+              'use server';
+              await signOut({
+                redirectTo: '/login',
+              });
+            }}
+            className="w-full"
+          >
+            <Button className="w-full">Logout</Button>
+          </form>
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
